Reorganize ReplState logic for cli/wasm compat
