+++ crates/repl_eval/src/gen.rs
@@ -50,7 +50,6 @@ pub fn compile_to_mono<'a, 'i, I: Iterator<Item = &'i str>>(
     defs: I,
     expr: &str,
     target_info: TargetInfo,
-    function_kind: FunctionKind,
     palette: Palette,
 ) -> (Option<MonomorphizedModule<'a>>, Problems) {
     let filename = PathBuf::from("");
@@ -64,7 +63,7 @@ pub fn compile_to_mono<'a, 'i, I: Iterator<Item = &'i str>>(
         RocCacheDir::Persistent(cache::roc_cache_dir().as_path()),
         LoadConfig {
             target_info,
-            function_kind,
+            function_kind: FunctionKind::LambdaSet,
             render: roc_reporting::report::RenderTarget::ColorTerminal,
             palette,
             threading: Threading::Single,
